# gen/http/membership_service/server/encode_decode.go
// Code generated by goa v3.22.6, DO NOT EDIT.
//
// membership-service HTTP server encoders and decoders
//
// Command:
// $ goa gen
// github.com/linuxfoundation/lfx-v2-member-service/cmd/member-api/design

package server

import (
	"context"
	"errors"
	"net/http"
	"strconv"
	"strings"

	membershipservice "github.com/linuxfoundation/lfx-v2-member-service/gen/membership_service"
	goahttp "goa.design/goa/v3/http"
	goa "goa.design/goa/v3/pkg"
)

// EncodeListMembershipsResponse returns an encoder for responses returned by
// the membership-service list-memberships endpoint.
func EncodeListMembershipsResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*membershipservice.ListMembershipsResult)
		enc := encoder(ctx, w)
		body := NewListMembershipsResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeListMembershipsRequest returns a decoder for requests sent to the
// membership-service list-memberships endpoint.
func DecodeListMembershipsRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*membershipservice.ListMembershipsPayload, error) {
	return func(r *http.Request) (*membershipservice.ListMembershipsPayload, error) {
		var (
			version     *string
			pageSize    int
			offset      int
			filter      *string
			bearerToken *string
			err         error
		)
		qp := r.URL.Query()
		versionRaw := qp.Get("v")
		if versionRaw != "" {
			version = &versionRaw
		}
		if version != nil {
			if !(*version == "1") {
				err = goa.MergeErrors(err, goa.InvalidEnumValueError("version", *version, []any{"1"}))
			}
		}
		{
			pageSizeRaw := qp.Get("pageSize")
			if pageSizeRaw == "" {
				pageSize = 25
			} else {
				v, err2 := strconv.ParseInt(pageSizeRaw, 10, strconv.IntSize)
				if err2 != nil {
					err = goa.MergeErrors(err, goa.InvalidFieldTypeError("pageSize", pageSizeRaw, "integer"))
				}
				pageSize = int(v)
			}
		}
		if pageSize < 1 {
			err = goa.MergeErrors(err, goa.InvalidRangeError("pageSize", pageSize, 1, true))
		}
		if pageSize > 100 {
			err = goa.MergeErrors(err, goa.InvalidRangeError("pageSize", pageSize, 100, false))
		}
		{
			offsetRaw := qp.Get("offset")
			if offsetRaw != "" {
				v, err2 := strconv.ParseInt(offsetRaw, 10, strconv.IntSize)
				if err2 != nil {
					err = goa.MergeErrors(err, goa.InvalidFieldTypeError("offset", offsetRaw, "integer"))
				}
				offset = int(v)
			}
		}
		if offset < 0 {
			err = goa.MergeErrors(err, goa.InvalidRangeError("offset", offset, 0, true))
		}
		filterRaw := qp.Get("filter")
		if filterRaw != "" {
			filter = &filterRaw
		}
		bearerTokenRaw := r.Header.Get("Authorization")
		if bearerTokenRaw != "" {
			bearerToken = &bearerTokenRaw
		}
		if err != nil {
			return nil, err
		}
		payload := NewListMembershipsPayload(version, pageSize, offset, filter, bearerToken)
		if payload.BearerToken != nil {
			if strings.Contains(*payload.BearerToken, " ") {
				// Remove authorization scheme prefix (e.g. "Bearer")
				cred := strings.SplitN(*payload.BearerToken, " ", 2)[1]
				payload.BearerToken = &cred
			}
		}

		return payload, nil
	}
}

// EncodeListMembershipsError returns an encoder for errors returned by the
// list-memberships membership-service endpoint.
func EncodeListMembershipsError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(ctx context.Context, err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		var en goa.GoaErrorNamer
		if !errors.As(v, &en) {
			return encodeError(ctx, w, v)
		}
		switch en.GoaErrorName() {
		case "BadRequest":
			var res *membershipservice.BadRequestError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewListMembershipsBadRequestResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusBadRequest)
			return enc.Encode(body)
		case "InternalServerError":
			var res *membershipservice.InternalServerError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewListMembershipsInternalServerErrorResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusInternalServerError)
			return enc.Encode(body)
		case "ServiceUnavailable":
			var res *membershipservice.ServiceUnavailableError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewListMembershipsServiceUnavailableResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusServiceUnavailable)
			return enc.Encode(body)
		default:
			return encodeError(ctx, w, v)
		}
	}
}

// EncodeGetMembershipResponse returns an encoder for responses returned by the
// membership-service get-membership endpoint.
func EncodeGetMembershipResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*membershipservice.GetMembershipResult)
		enc := encoder(ctx, w)
		body := NewGetMembershipResponseBody(res)
		if res.Etag != nil {
			w.Header().Set("Etag", *res.Etag)
		}
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeGetMembershipRequest returns a decoder for requests sent to the
// membership-service get-membership endpoint.
func DecodeGetMembershipRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*membershipservice.GetMembershipPayload, error) {
	return func(r *http.Request) (*membershipservice.GetMembershipPayload, error) {
		var (
			uid         string
			version     *string
			bearerToken *string
			err         error

			params = mux.Vars(r)
		)
		uid = params["uid"]
		err = goa.MergeErrors(err, goa.ValidateFormat("uid", uid, goa.FormatUUID))
		versionRaw := r.URL.Query().Get("v")
		if versionRaw != "" {
			version = &versionRaw
		}
		if version != nil {
			if !(*version == "1") {
				err = goa.MergeErrors(err, goa.InvalidEnumValueError("version", *version, []any{"1"}))
			}
		}
		bearerTokenRaw := r.Header.Get("Authorization")
		if bearerTokenRaw != "" {
			bearerToken = &bearerTokenRaw
		}
		if err != nil {
			return nil, err
		}
		payload := NewGetMembershipPayload(uid, version, bearerToken)
		if payload.BearerToken != nil {
			if strings.Contains(*payload.BearerToken, " ") {
				// Remove authorization scheme prefix (e.g. "Bearer")
				cred := strings.SplitN(*payload.BearerToken, " ", 2)[1]
				payload.BearerToken = &cred
			}
		}

		return payload, nil
	}
}

// EncodeGetMembershipError returns an encoder for errors returned by the
// get-membership membership-service endpoint.
func EncodeGetMembershipError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(ctx context.Context, err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		var en goa.GoaErrorNamer
		if !errors.As(v, &en) {
			return encodeError(ctx, w, v)
		}
		switch en.GoaErrorName() {
		case "InternalServerError":
			var res *membershipservice.InternalServerError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewGetMembershipInternalServerErrorResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusInternalServerError)
			return enc.Encode(body)
		case "NotFound":
			var res *membershipservice.NotFoundError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewGetMembershipNotFoundResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusNotFound)
			return enc.Encode(body)
		case "ServiceUnavailable":
			var res *membershipservice.ServiceUnavailableError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewGetMembershipServiceUnavailableResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusServiceUnavailable)
			return enc.Encode(body)
		default:
			return encodeError(ctx, w, v)
		}
	}
}

// EncodeListMembershipContactsResponse returns an encoder for responses
// returned by the membership-service list-membership-contacts endpoint.
func EncodeListMembershipContactsResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*membershipservice.ListMembershipContactsResult)
		enc := encoder(ctx, w)
		body := NewListMembershipContactsResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeListMembershipContactsRequest returns a decoder for requests sent to
// the membership-service list-membership-contacts endpoint.
func DecodeListMembershipContactsRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*membershipservice.ListMembershipContactsPayload, error) {
	return func(r *http.Request) (*membershipservice.ListMembershipContactsPayload, error) {
		var (
			uid         string
			version     *string
			bearerToken *string
			err         error

			params = mux.Vars(r)
		)
		uid = params["uid"]
		err = goa.MergeErrors(err, goa.ValidateFormat("uid", uid, goa.FormatUUID))
		versionRaw := r.URL.Query().Get("v")
		if versionRaw != "" {
			version = &versionRaw
		}
		if version != nil {
			if !(*version == "1") {
				err = goa.MergeErrors(err, goa.InvalidEnumValueError("version", *version, []any{"1"}))
			}
		}
		bearerTokenRaw := r.Header.Get("Authorization")
		if bearerTokenRaw != "" {
			bearerToken = &bearerTokenRaw
		}
		if err != nil {
			return nil, err
		}
		payload := NewListMembershipContactsPayload(uid, version, bearerToken)
		if payload.BearerToken != nil {
			if strings.Contains(*payload.BearerToken, " ") {
				// Remove authorization scheme prefix (e.g. "Bearer")
				cred := strings.SplitN(*payload.BearerToken, " ", 2)[1]
				payload.BearerToken = &cred
			}
		}

		return payload, nil
	}
}

// EncodeListMembershipContactsError returns an encoder for errors returned by
// the list-membership-contacts membership-service endpoint.
func EncodeListMembershipContactsError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(ctx context.Context, err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		var en goa.GoaErrorNamer
		if !errors.As(v, &en) {
			return encodeError(ctx, w, v)
		}
		switch en.GoaErrorName() {
		case "InternalServerError":
			var res *membershipservice.InternalServerError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewListMembershipContactsInternalServerErrorResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusInternalServerError)
			return enc.Encode(body)
		case "NotFound":
			var res *membershipservice.NotFoundError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewListMembershipContactsNotFoundResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusNotFound)
			return enc.Encode(body)
		case "ServiceUnavailable":
			var res *membershipservice.ServiceUnavailableError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewListMembershipContactsServiceUnavailableResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusServiceUnavailable)
			return enc.Encode(body)
		default:
			return encodeError(ctx, w, v)
		}
	}
}

// EncodeReadyzResponse returns an encoder for responses returned by the
// membership-service readyz endpoint.
func EncodeReadyzResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.([]byte)
		ctx = context.WithValue(ctx, goahttp.ContentTypeKey, "text/plain")
		enc := encoder(ctx, w)
		body := res
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// EncodeReadyzError returns an encoder for errors returned by the readyz
// membership-service endpoint.
func EncodeReadyzError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(ctx context.Context, err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		var en goa.GoaErrorNamer
		if !errors.As(v, &en) {
			return encodeError(ctx, w, v)
		}
		switch en.GoaErrorName() {
		case "ServiceUnavailable":
			var res *membershipservice.ServiceUnavailableError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewReadyzServiceUnavailableResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusServiceUnavailable)
			return enc.Encode(body)
		default:
			return encodeError(ctx, w, v)
		}
	}
}

// EncodeLivezResponse returns an encoder for responses returned by the
// membership-service livez endpoint.
func EncodeLivezResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.([]byte)
		ctx = context.WithValue(ctx, goahttp.ContentTypeKey, "text/plain")
		enc := encoder(ctx, w)
		body := res
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// marshalMembershipserviceMembershipResponseToMembershipResponseResponseBody
// builds a value of type *MembershipResponseResponseBody from a value of type
// *membershipservice.MembershipResponse.
func marshalMembershipserviceMembershipResponseToMembershipResponseResponseBody(v *membershipservice.MembershipResponse) *MembershipResponseResponseBody {
	res := &MembershipResponseResponseBody{
		UID:              v.UID,
		Name:             v.Name,
		Status:           v.Status,
		Year:             v.Year,
		Tier:             v.Tier,
		MembershipType:   v.MembershipType,
		AutoRenew:        v.AutoRenew,
		RenewalType:      v.RenewalType,
		Price:            v.Price,
		AnnualFullPrice:  v.AnnualFullPrice,
		PaymentFrequency: v.PaymentFrequency,
		PaymentTerms:     v.PaymentTerms,
		AgreementDate:    v.AgreementDate,
		PurchaseDate:     v.PurchaseDate,
		StartDate:        v.StartDate,
		EndDate:          v.EndDate,
		CreatedAt:        v.CreatedAt,
		UpdatedAt:        v.UpdatedAt,
	}
	if v.Account != nil {
		res.Account = marshalMembershipserviceAccountTypeToAccountTypeResponseBody(v.Account)
	}
	if v.Contact != nil {
		res.Contact = marshalMembershipserviceContactTypeToContactTypeResponseBody(v.Contact)
	}
	if v.Product != nil {
		res.Product = marshalMembershipserviceProductTypeToProductTypeResponseBody(v.Product)
	}
	if v.Project != nil {
		res.Project = marshalMembershipserviceProjectTypeToProjectTypeResponseBody(v.Project)
	}

	return res
}

// marshalMembershipserviceAccountTypeToAccountTypeResponseBody builds a value
// of type *AccountTypeResponseBody from a value of type
// *membershipservice.AccountType.
func marshalMembershipserviceAccountTypeToAccountTypeResponseBody(v *membershipservice.AccountType) *AccountTypeResponseBody {
	if v == nil {
		return nil
	}
	res := &AccountTypeResponseBody{
		ID:      v.ID,
		Name:    v.Name,
		LogoURL: v.LogoURL,
		Website: v.Website,
	}

	return res
}

// marshalMembershipserviceContactTypeToContactTypeResponseBody builds a value
// of type *ContactTypeResponseBody from a value of type
// *membershipservice.ContactType.
func marshalMembershipserviceContactTypeToContactTypeResponseBody(v *membershipservice.ContactType) *ContactTypeResponseBody {
	if v == nil {
		return nil
	}
	res := &ContactTypeResponseBody{
		ID:        v.ID,
		FirstName: v.FirstName,
		LastName:  v.LastName,
		Email:     v.Email,
		Title:     v.Title,
	}

	return res
}

// marshalMembershipserviceProductTypeToProductTypeResponseBody builds a value
// of type *ProductTypeResponseBody from a value of type
// *membershipservice.ProductType.
func marshalMembershipserviceProductTypeToProductTypeResponseBody(v *membershipservice.ProductType) *ProductTypeResponseBody {
	if v == nil {
		return nil
	}
	res := &ProductTypeResponseBody{
		ID:     v.ID,
		Name:   v.Name,
		Family: v.Family,
		Type:   v.Type,
	}

	return res
}

// marshalMembershipserviceProjectTypeToProjectTypeResponseBody builds a value
// of type *ProjectTypeResponseBody from a value of type
// *membershipservice.ProjectType.
func marshalMembershipserviceProjectTypeToProjectTypeResponseBody(v *membershipservice.ProjectType) *ProjectTypeResponseBody {
	if v == nil {
		return nil
	}
	res := &ProjectTypeResponseBody{
		ID:      v.ID,
		Name:    v.Name,
		LogoURL: v.LogoURL,
		Slug:    v.Slug,
		Status:  v.Status,
	}

	return res
}

// marshalMembershipserviceListMetadataToListMetadataResponseBody builds a
// value of type *ListMetadataResponseBody from a value of type
// *membershipservice.ListMetadata.
func marshalMembershipserviceListMetadataToListMetadataResponseBody(v *membershipservice.ListMetadata) *ListMetadataResponseBody {
	res := &ListMetadataResponseBody{
		TotalSize: v.TotalSize,
		PageSize:  v.PageSize,
		Offset:    v.Offset,
	}

	return res
}

// marshalMembershipserviceKeyContactResponseToKeyContactResponseResponseBody
// builds a value of type *KeyContactResponseResponseBody from a value of type
// *membershipservice.KeyContactResponse.
func marshalMembershipserviceKeyContactResponseToKeyContactResponseResponseBody(v *membershipservice.KeyContactResponse) *KeyContactResponseResponseBody {
	res := &KeyContactResponseResponseBody{
		UID:            v.UID,
		MembershipUID:  v.MembershipUID,
		Role:           v.Role,
		Status:         v.Status,
		BoardMember:    v.BoardMember,
		PrimaryContact: v.PrimaryContact,
		CreatedAt:      v.CreatedAt,
		UpdatedAt:      v.UpdatedAt,
	}
	if v.Contact != nil {
		res.Contact = marshalMembershipserviceContactTypeToContactTypeResponseBody(v.Contact)
	}
	if v.Project != nil {
		res.Project = marshalMembershipserviceProjectTypeToProjectTypeResponseBody(v.Project)
	}
	if v.Organization != nil {
		res.Organization = marshalMembershipserviceOrganizationTypeToOrganizationTypeResponseBody(v.Organization)
	}

	return res
}

// marshalMembershipserviceOrganizationTypeToOrganizationTypeResponseBody
// builds a value of type *OrganizationTypeResponseBody from a value of type
// *membershipservice.OrganizationType.
func marshalMembershipserviceOrganizationTypeToOrganizationTypeResponseBody(v *membershipservice.OrganizationType) *OrganizationTypeResponseBody {
	if v == nil {
		return nil
	}
	res := &OrganizationTypeResponseBody{
		ID:      v.ID,
		Name:    v.Name,
		LogoURL: v.LogoURL,
		Website: v.Website,
	}

	return res
}