# gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Membership Management Service
    version: 0.0.1
host: localhost:80
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
    /memberships:
        get:
            tags:
                - membership-service
            summary: list-memberships membership-service
            description: List memberships with pagination and filtering
            operationId: membership-service#list-memberships
            parameters:
                - name: v
                  in: query
                  description: Version of the API
                  required: false
                  type: string
                  enum:
                    - "1"
                - name: pageSize
                  in: query
                  description: Number of items per page
                  required: false
                  type: integer
                  default: 25
                  maximum: 100
                  minimum: 1
                - name: offset
                  in: query
                  description: Offset into the total list
                  required: false
                  type: integer
                  default: 0
                  minimum: 0
                - name: filter
                  in: query
                  description: Filter expression (key=value pairs separated by semicolons)
                  required: false
                  type: string
                - name: Authorization
                  in: header
                  description: JWT token issued by Heimdall
                  required: false
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/MembershipServiceListMembershipsResponseBody'
                        required:
                            - memberships
                            - metadata
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/BadRequestError'
                        required:
                            - message
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/InternalServerError'
                        required:
                            - message
                "503":
                    description: Service Unavailable response.
                    schema:
                        $ref: '#/definitions/ServiceUnavailableError'
                        required:
                            - message
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
    /memberships/{uid}:
        get:
            tags:
                - membership-service
            summary: get-membership membership-service
            description: Get a specific membership by UID
            operationId: membership-service#get-membership
            parameters:
                - name: v
                  in: query
                  description: Version of the API
                  required: false
                  type: string
                  enum:
                    - "1"
                - name: uid
                  in: path
                  description: Membership UID
                  required: true
                  type: string
                  format: uuid
                - name: Authorization
                  in: header
                  description: JWT token issued by Heimdall
                  required: false
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/MembershipServiceGetMembershipResponseBody'
                    headers:
                        ETag:
                            description: ETag header value
                            type: string
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/NotFoundError'
                        required:
                            - message
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/InternalServerError'
                        required:
                            - message
                "503":
                    description: Service Unavailable response.
                    schema:
                        $ref: '#/definitions/ServiceUnavailableError'
                        required:
                            - message
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
    /memberships/{uid}/contacts:
        get:
            tags:
                - membership-service
            summary: list-membership-contacts membership-service
            description: Get key contacts for a specific membership
            operationId: membership-service#list-membership-contacts
            parameters:
                - name: v
                  in: query
                  description: Version of the API
                  required: false
                  type: string
                  enum:
                    - "1"
                - name: uid
                  in: path
                  description: Membership UID
                  required: true
                  type: string
                  format: uuid
                - name: Authorization
                  in: header
                  description: JWT token issued by Heimdall
                  required: false
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/MembershipServiceListMembershipContactsResponseBody'
                        required:
                            - contacts
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/NotFoundError'
                        required:
                            - message
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/InternalServerError'
                        required:
                            - message
                "503":
                    description: Service Unavailable response.
                    schema:
                        $ref: '#/definitions/ServiceUnavailableError'
                        required:
                            - message
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
definitions:
    AccountType:
        title: AccountType
        type: object
        properties:
            id:
                type: string
                description: Account ID
                example: acc-123
            logo_url:
                type: string
                description: Account logo URL
                example: https://example.com/logo.png
            name:
                type: string
                description: Account name
                example: Example Corp
            website:
                type: string
                description: Account website
                example: https://example.com
        example:
            id: acc-123
            logo_url: https://example.com/logo.png
            name: Example Corp
            website: https://example.com
    BadRequestError:
        title: BadRequestError
        type: object
        properties:
            message:
                type: string
                description: Error message
                example: The request was invalid.
        description: Bad request
        example:
            message: The request was invalid.
        required:
            - message
    ContactType:
        title: ContactType
        type: object
        properties:
            email:
                type: string
                description: Email address
                example: john.doe@example.com
            first_name:
                type: string
                description: First name
                example: John
            id:
                type: string
                description: Contact ID
                example: con-123
            last_name:
                type: string
                description: Last name
                example: Doe
            title:
                type: string
                description: Job title
                example: CTO
        example:
            email: john.doe@example.com
            first_name: John
            id: con-123
            last_name: Doe
            title: CTO
    InternalServerError:
        title: InternalServerError
        type: object
        properties:
            message:
                type: string
                description: Error message
                example: An internal server error occurred.
        description: Internal server error
        example:
            message: An internal server error occurred.
        required:
            - message
...
    MembershipResponse:
        title: MembershipResponse
        type: object
        properties:
            account:
                $ref: '#/definitions/AccountType'
            agreement_date:
                type: string
                description: Agreement date
                example: "2025-01-01T00:00:00Z"
            annual_full_price:
                type: number
                description: Annual full price
                example: 50000
                format: double
            auto_renew:
                type: boolean
                description: Whether auto-renew is enabled
                example: true
            contact:
                $ref: '#/definitions/ContactType'
            created_at:
                type: string
                description: Creation timestamp
                example: "2025-01-01T00:00:00Z"
                format: date-time
            end_date:
                type: string
                description: End date
                example: "2025-12-31T23:59:59Z"
            membership_type:
                type: string
                description: Membership type
                example: Corporate
            name:
                type: string
                description: Membership name
                example: Gold Membership - Example Corp
            payment_frequency:
                type: string
                description: Payment frequency
                example: Annual
            payment_terms:
                type: string
                description: Payment terms
                example: Net 30
            price:
                type: number
                description: Membership price
                example: 50000
                format: double
            product:
                $ref: '#/definitions/ProductType'
            project:
                $ref: '#/definitions/ProjectType'
            purchase_date:
                type: string
                description: Purchase date
                example: "2025-01-01T00:00:00Z"
            renewal_type:
                type: string
                description: Renewal type
                example: Annual
            start_date:
                type: string
                description: Start date
                example: "2025-01-01T00:00:00Z"
            status:
                type: string
                description: Membership status
                example: Active
            tier:
                type: string
                description: Membership tier
                example: Gold
            uid:
                type: string
                description: Membership UID
                example: 7cad5a8d-19d0-41a4-81a6-043453daf9ee
                format: uuid
            updated_at:
                type: string
                description: Last update timestamp
                example: "2025-06-01T12:00:00Z"
                format: date-time
            year:
                type: string
                description: Membership year
                example: "2025"
        description: A membership resource
        example:
            account:
                id: acc-123
                logo_url: https://example.com/logo.png
                name: Example Corp
                website: https://example.com
            agreement_date: "2025-01-01T00:00:00Z"
            annual_full_price: 50000
            auto_renew: true
            contact:
                email: john.doe@example.com
                first_name: John
                id: con-123
                last_name: Doe
                title: CTO
            created_at: "2025-01-01T00:00:00Z"
            end_date: "2025-12-31T23:59:59Z"
            membership_type: Corporate
            name: Gold Membership - Example Corp
            payment_frequency: Annual
            payment_terms: Net 30
            price: 50000
            product:
                family: Membership
                id: prod-123
                name: Gold Membership
                type: Corporate
            project:
                id: proj-123
                logo_url: https://example.com/project-logo.png
                name: Linux Foundation
                slug: linux-foundation
                status: Active
            purchase_date: "2025-01-01T00:00:00Z"
            renewal_type: Annual
            start_date: "2025-01-01T00:00:00Z"
            status: Active
            tier: Gold
            uid: 7cad5a8d-19d0-41a4-81a6-043453daf9ee
            updated_at: "2025-06-01T12:00:00Z"
            year: "2025"
    MembershipServiceGetMembershipResponseBody:
        title: MembershipServiceGetMembershipResponseBody
        $ref: '#/definitions/MembershipResponse'
...
    NotFoundError:
        title: NotFoundError
        type: object
        properties:
            message:
                type: string
                description: Error message
                example: The resource was not found.
        description: Resource not found
        example:
            message: The resource was not found.
        required:
            - message
...
    ServiceUnavailableError:
        title: ServiceUnavailableError
        type: object
        properties:
            message:
                type: string
                description: Error message
                example: The service is unavailable.
        description: Service unavailable
        example:
            message: The service is unavailable.
        required:
            - message
securityDefinitions:
    jwt_header_Authorization:
        type: apiKey
        description: Heimdall authorization
        name: Authorization
        in: header